requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
pandas>=1.3
numpy>=1.21
matplotlib>=3.5
scikit-learn>=1.0
pytest>=7.0
imageio>=2.13
imageio-ffmpeg>=0.4
Pillow>=9.0
